Rename to hyper-react (1/N)
* gemspace renamed
* top level file renamed
