+++ lib/generators/reactive_ruby/test_app/templates/assets/javascripts/components.rb
@@ -1,4 +1,4 @@
 require 'opal'
 require 'react/react-source'
-require 'reactrb'
+require 'hyper-react'
 require_tree './components'
