+++ lib/generators/reactive_ruby/test_app/templates/test_application.rb.erb
@@ -7,7 +7,7 @@ require File.expand_path('../boot', __FILE__)
 Bundler.require(*Rails.groups(assets: %w(development test)))
 
 require 'opal-rails'
-require 'reactrb'
+require 'hyper-react'
 
 <%= application_definition %>
 <% end %>
