+++ lib/hyper-react.rb
