+++ spec/spec_helper.rb
@@ -16,7 +16,7 @@ end
 if RUBY_ENGINE == 'opal'
   require File.expand_path('../vendor/jquery-2.2.4.min', __FILE__)
   require 'react/react-source'
-  require 'reactrb'
+  require 'hyper-react'
   require 'react/test/rspec'
 
   require File.expand_path('../support/react/spec_helpers', __FILE__)
